Update NativePropertiesAdmin in C#
